pdf-py/requirements.txt
flask==3.0.0
flask-cors==3.0.10
werkzeug==3.0.1
pdfplumber==0.11.4
pymupdf==1.24.1
pdfminer.six==20231228
# pdfminer.six==20240706
snakeviz